+++ src/main/java/sk/stuba/fei/uim/oop/tiles/StartEnd.java
@@ -30,7 +30,7 @@ public class StartEnd extends Tile{
             this.highlight =false;
         }
         g2d.setStroke(new BasicStroke(4));
-        g2d.drawRect(0, (int)this.getHeight()/2-this.getHeight()/6, this.getWidth()/3*2, (int)this.getWidth()/3);
+        g2d.drawRect(0, this.getHeight() /2-this.getHeight()/6, this.getWidth()/3*2, this.getWidth() /3);
         g2d.drawOval((int) (this.getWidth()/2 - this.getWidth() * 0.35), (int) (this.getHeight()/2 - this.getHeight() * 0.35), (int) (this.getWidth() * 0.7), (int) (this.getHeight() * 0.7));
         g2d.setStroke(new BasicStroke(1));
         if(this.water || this.start){
@@ -39,7 +39,7 @@ public class StartEnd extends Tile{
         else {
             g2d.setColor(Color.GRAY);
         }
-        g2d.fillRect(0, (int)this.getHeight()/2-this.getHeight()/6, this.getWidth()/3*2, (int)this.getWidth()/3);
+        g2d.fillRect(0, this.getHeight() /2-this.getHeight()/6, this.getWidth()/3*2, this.getWidth() /3);
         g2d.fillOval((int) (this.getWidth()/2 - this.getWidth() * 0.35), (int) (this.getHeight()/2 - this.getHeight() * 0.35), (int) (this.getWidth() * 0.7), (int) (this.getHeight() * 0.7));
         g2d.setColor(Color.BLACK);
         g2d.dispose();
